Store generated css back in src/main/resources
* Need to bind to an earlier Maven lifecycle phase to get the assets
into the jar file
* Once that is done, it makes more sense (to me anyway) to store these
files in src/main/resources and in git.  That way they are always
available and users that clone the repository and just run the app
without building a jar file will see the styles working. Since git
only marks a file as dirty if it changes it doesn't really matter that
the files are overwritten on every build.