Append trailing '/' to GSGs instead of '/content'
Prior to this commit all GSG URLs were suffixed with /content in order
to ensure that relative image paths would resolve properly. The
GettingStartedController was responsible for redirecting requests for
/guide-name to /guide-name/content.

This commit cleans up these paths as much as possible by appending a
simple trailing slash instead of the more verbose '/content'. Redirect
logic for rewriting requests for /guide-name to /guide-name/ is now
handled by the UrlRewriteFilter.

 - Rewrite requests for /guides/gs/name => /guides/gs/name/
 - Update all templates, tests, etc to reflect new URLs
 - Individual GSGs have also been updated to reflect new URLs [#54822628]

[#Fixes 52214321]