Disable indexer deployment and legacy url mapping
This commit comments out indexer deployment in blue-green deploys and in
the development manifest in order to speed up deployment and avoid false
cf push failures due to timeouts that frequently occur when trying to
start the indexer app.

This commit also comments out the mapping of legacy urls such as
springsource.org, interface21.com etc when performing production
deployments, once again for the benefit of speed.

This is a temporary arrangement, and it is expected that this commit
will be reverted as soon as feasible.