Extract legacy site redirects out of main site
This commit removes "legacy redirects", i.e. all those that deal with
older, non-spring.io domans such as springsource.org, interface21.com,
etc, and leaves in place those redirects that are specific to the
spring.io application, e.g /guides/tutorials => /guides#tutorials and so
on.

The legacy redirects are being ported to a separate application.