Support tunneling arbitrary HTML through GH Markdown API
GitHub's Markdown rendering API [1] strips all HTML tags save those on
its whitelist. Tags such as <iframe> and <embed> are not on that
whitelist. This is easily demonstrated with the following example:

    echo '<a/><z/>' | curl -XPOST https://api.github.com/markdown/raw -H"Content-Type: text/plain" -d @-

Notice in the result that the whitelisted tag <a> is preserved, while
the unknown tag <z> is stripped.

This situation presents a problem for our blog, because posts are
written in Markdown and then processed through GitHub's API to render
them as HTML. The practical effect of this situation is that blog
authors have not to date been able to embed YouTube videos and other
media on our blog.

This commit works around the limitations of the GitHub Markdown API by
specifying a format for 'encoding' arbitrary HTML tags in the source of
a blog post and supporting decoding of those tags after the post has been
processed through the GH API. In effect, this allows blog authors to
'tunnel' any content they may need through the API in a way that ensures
GH will not mangle or strip it out. The encoding format is as follows:

    <pre>!{iframe src="//www.youtube.com/embed/jplkJIHPGos"}{/iframe}</pre>

 - The <pre> tags ensure that GH does not try to automatically wrap any
   URLs in <a> tags.

 - The ! following the opening <pre> tag allows the parser to
   unambiguously determine that this is a candidate for decoding.

 - The use of curly braces ('{') instead of angle brackets ('<') ensures
   that GH does not strip the tag in question ('iframe') in the example
   above.

The example above, when decoded, renders as follows:

    <iframe src="//www.youtube.com/embed/jplkJIHPGos"></iframe>

The process, then, for including a YouTube video in a blog post is:

 1. Begin with the iframe HTML as provided by YouTube
 2. Replace angle brackets with curly brances
 3. Wrap everything in <pre>!...</pre>
 4. Include within the blog post wherever appropriate

Notes:

 - Any number of encoded <pre>!...</pre> sequences may be included in a
   given blog post.

 - Put each <pre>!...</pre> sequence on its own line for best results.

[Finishes #57789824]
[1]: http://developer.github.com/v3/markdown/