Migrate from Liquibase to Flyway, from HSQL to H2
This change replaces Liquibase[1] with Flyway[2] for database schema
migrations for the following reasons:

1. To avoid DB lock contention issues at startup

We dealt with a number of production outages due to the fact that
Liquibase caused (or otherwise contributed to) lock contention issues
at startup (i.e. when Liquibase needs to lock tables in order to apply
changes to the DB) that resulted in DB connection pool starvation. This
behavior was never isolated and repeated in a test environment, but was
observed several times in practice in production. The workaround has been
to reduce the instance count in production from 4 to 1, but this has its
own negative effects (no redundancy in the case of an app crashing).

So this move to Flyway is in some ways an experiment in which we will
see over time whether it too can result in these locking / starvation
issues. Testing the application under load in an 8-instance environment
suggests that it will not, but only sustained time in production will
tell for sure.

2. Flyway's plain-SQL approach is clearer and more suitable to our needs

Liquibase has a powerful DSL that acts as an abstraction layer over SQL,
making it simple to develop in memory against HSQL while deploying to
production against Postgres. However, it comes at the cost of needing
to understand that DSL (which in practice means looking things up in
Liquibase's documentation each time a change needs to be made). This is
a minor hassle for project committers, but is more important when
considering this application's purpose as a reference app; plain SQL is
immediately clear to a larger percentage of the audience than
Liquibase's DSL would be, therefore Flyway's approach is preferred.

3. Flyway is (or appears to be) faster than Liquibase at startup time

Liquibase initialization took approximately 3 seconds on an MBP against
an in-memory DB. Flyway initialization takes less than 1 second. This is
significant given that it takes overall startup time from ~12 seconds to
~8 seconds, i.e. more efficient when in a tight loop of code changes and
restarts. Note that this comparison isn't perfectly apples-to-apples, as
the Liquibase arrongement had 3 changesets whereas there is now just one
consolidated changeset under Flyway.

This commit also migrates from HSQL to H2 as the in-memory database used
for development and automated testing. The primary reason is because H2
offers an explicit "PostgreSQL compatibility mode", making it feasible
to configure Flyway with a unified set of SQL files, as opposed to
maintaining two copies, one for the in-memory DB and one for the
production DB. HSQL does not have this option. This change was quite
straightforward to perform - H2 was effectively a drop-in replacement,
and side-by-side comparisons of executing the test suites demonstrates
little or no difference with regard to performance.

[Finishes #57709992]
[1]: http://www.liquibase.org
[2]: http://flywaydb.org