Add script to flip (www.)spring.io between spaces
This script is useful when blue/green deployment within a single space
is not enough, e.g. in cases where all traffic to the database must be
stopped in order to fundamentally rebuild it (as was the case in the
previous commits).

This approach requires a few seconds of production downtime, but keeps
it to an absolute minimum, on the order of 30 seconds total.

[#57709992]