Add basic buster setup for node-based unit tests.
This adds a simple buster config, an npm test command to point to the
right spot to run buster, and a simple example unit test to show
how to write tests using buster's describe/expect style.  I created
a simple helper that exposes the basic spec style dsl globally,
including expect().  We'll have to decide whether we want to stick
to that path (globalizing the spec dsl) or not.