Add basic buster selenium webdriverjs setup
This adds the webdriverjs client wrapped in a custom buster extension
that injects it into scenario tests (but not unit tests), and adds new
npm scripts for running against phantomjs for now.

There is apparently a bug in buster whereby exposing the
describe-style globals before extensions are initialized causes the
`this` context of tests to get messed up.  Consequently, I had to
switch to exposing them in the individual test files.  We need to
find a better way until the buster problem is fixed.