Add initial Guides functional tests
Moved `example-spec` to `main-spec` as it is a valid test for the main
sagan page.

Ditched webdriverjs in favor of wd, which seems to have a much more
extensive and flexible API.  It also returns promises, so we can use
when.js's promise combinators if necessary to do more complex tests.

Updated README with a better selenium-server download link, and added
instructions for turning on Buster's specification style reporter if
you want more output while tests run.