Add selenium tests for searching
The tests make sure the search widget is not visible initially on the
main page, and then walk through a typical search interaction: click
to open search, type something, submit, ensure results are shown (but
don't validate the actual results), and that the search widget is open
by default and prefilled on the results page.

I ran into a problem with phantomjs, which, after much research and
blood/sweat/tears, I've punted on for now.  It seems that in
phantomjs, the element you click to open the search never passes
Selenium's heuristics for being visible--no matter how long you wait
after making it visible (by adding the js-show class). Selenium
refuses to click an element it considers not to be visible.  Thus,
it's impossible to test search in a way that simulates real user
interactions in phantom--we might be able to fake it by forcibly
adding a class to the search dropdown rather than relying on the
click, but that seems like a total hack.

So, for now, I've added a new buster config group, and an associated
npm script to run selenium tests in firefox: `npm run firefox`.  See
sagan-js/test/README.md for more info.

Updated to latest wd selenium driver version.

[Fixes #58259306]