Use data-* attributes to identify required page capabilities
In an ideal world, we would componentize the html and css alongside
the JS.  However, we need a shorter term solution that doesn't require
getting into such a lengthy task.  The approach we decided on is to
tag each page, via its top-level html element, to identify what
coarse-grained capabilities it needs.  For example: search, or a
filterable list, or info popups.  That allows us to scan the html
element (which can be done *before* DOMReady) and initialize only
those capabilities, rather than attempting to initialize all possible
capabilities.

In main.js the capabilities are laid out in a simple hash.  Many, but
not all of the individual capabilities have been factored out into
separate modules yet, but this gives us the structure to do so.