Restrict GitHub OAuth Scope
Resolve #30 by changing GitHub OAuth scope from the private data-hungry
'user' scope to the less greedy (and lesser-known) 'read:org' scope.

The app only needs to inquire as to whether the current principal is a
member of the correct GitHub team. The latter scope allows this without
requiring the user to give up access to private repos and other
sensitive data.