Revert "Restore 3-column layout on /guides page"
This reverts commit 572a5cf2fa770f937c5e74cf633b535a2a1ba74b in order to
apply the same change to the actual source file for guide.css, which is
now in sagan-client/src/css/guide.css. The file modified by the commit
above is actually a product of the build process in sagan-client.

In the future, changes to .css and .js resources should always be done
first in sagan-client. Then run the following commands to propagate the
changes to sagan-site:

    $ npm run build
    $ npm run dump

When this is complete, there will be modifications in the files in
sagan-site, and you can then check them in.