Revert "Restore 3-column layout on /guides page"
This reverts commit 8407a6a5b60901146bbeb49b89949580e194a289 in order to
apply the same change to the actual source file for guide.css, which is
now in sagan-client/src/css/guide.css. The file modified by the commit
above is actually a product of the build process in sagan-client.

In the future, changes to .css and .js resources should always be done
first in sagan-client. Then run the following commands to propagate the
changes to sagan-site:

    $ npm run build
    $ npm run dump

When this is complete, there will be modifications in the files in
sagan-site, and you can then check them in.