Increase PermGen space when using `gradle run`
Prior to this commit, `gradle run` would use the default MaxPermSize
value of 64M on JDK7. In practice this is not enough to handle what is
assumed to be a significant amount of dynamic classloading by JRuby and
asciidoctor when processing our Asciidoc-based guides.

MaxPermSize is now set to 128M, which manual tests confirm is enough to
avoid these errors.