Simplify declaration of 'deploy' task
Previously, the 'deploy' task was declared using a Gradle 'task rule',
which is no longer necessary now that the name of the space is taken as
a -P parameter, as opposed to being part of a dynamic task name.

Also, mustRunAfter logic has been removed with regard to indexer's
'deploy' task always running after site's deploy task. This is already
enforced elsewhere, meaning that this logic is redundant.