Defend against changing project names in deploy.gradle
Previously, project name was compared to simple string, e.g.
'sagan-site'. Now, the name of the 'sagan-site' project is looked up
with a call to project(':sagan-site').name. While this seems redundant
at first glance, the advantage is that the project() call will fail fast
and explicitly if 'sagan-site' no longer exists as a project (or its
name has been changed).