Fix LazyInitializationException for Project
Prior to this commit, the list of releases for a given project was not
fetched from the database when using the "findAll" repository method.
Since OpenEntityManagerInViewFilter does not come into play for this,
the app would throw a LazyInitializationException when accessing the
list of releases for a given project.

This commit adds a new method in the ProjectDataRepository that also
fetches releases information in the same query and avoid later lazy
init exceptions.

Issue: