Resolves #275: Have guides use project metadata to make sidebar links
Currently, the sidebar links for projects wrangle the YAML front matter to project a link to the project page and name. This patch uses the ProjectMetadataService to instead use a proper site URL and name.

Tested by running the app, adding a new project for Grails/http://grails.org. Then I had a forked copy of gs-accessing-data-gorm with "grails" in the YAML projects. It worked great at rendering both Spring Boot and Grails in the sidebar.