Use gulp-bower-src plugin
Prior to this commit, the gulp build was using gulp-bower-files, which
is a plugin that fetches "main" files from bower components.
"main" files are the main point of entry for a given bower component,
and are not the complete list of required files for production.

This commit switches the build to gulp-bower-src, a gulp plugin that
fetches all files from components, except the ones matching patterns
listed in the "ignore" property of the component's bower.json.
Since many components don't have such an attribute, then this plugin
also reads "ignore" properties in the "override" part of the main
bower.json configuration file.

See: https://github.com/bclozel/gulp-bower-src/