Fix type mismatch errors from Postgres on blog update
The definition of the new POST_PUBLICSLUGALIASES tables as introduced
in #284 had a POST_ID field with type varchar. This should have been
integer from the start, since it is storing values of POST.ID, which
itself is serial (integer).

This varchar vs. integer type mismatch was succeeding under H2 (even in
postgres-compatibility mode), but failing in staging / production under
an actual Postgres installation.

This commit alters the column definition to type integer.