Instant reloading of static resources during dev
Prior to this commit, one had to restart the whole sagan-site app or
rely on gulp-watch scripts to see changes live in a locally running
sagan-site application.

This commit adds another resource location that serves static resources
from the `sagan-client` module directly. This feature is automatically
enabled when running the app with gradle:

    $ gradle sagan-site:bootRun

This commit also deletes previous gulp-watch tasks.