Improve error message for search errors
When running the Sagan app for the first time, the app is probably
not configured to use an existing ElasticSearch index. So the search
feature doesn't work properly.

This commit adds a specific error page that points to Sagan wiki pages,
thus providing insight on how to properly use an ElasticSearch instance
with the Sagan app.