Fixes invalid value for ddl-auto
Prior to this commit the property spring.jpa.hibernate.ddl-auto was set to 'false', not a valid value for that property - and a warning was issued by Hibernate. The wanted behaviour is to have "no automatic DDL actions", so the correct is to set it to the (undocumented) value 'none'.