Update tests to reflect GitHub's new path to raw content
The dynamic sidebar tests use macros, which are located on github via
raw URLs. When they changed their path, it didn't cause any failures
while also slowing down the tests (see #369).

This commit updates the tests to use the new paths and also added a
check to verify that macro content was actually loaded. Both should
cause a more recognizable failure if github makes similar edits.

The impact is that all the guides need to be updated to reflect the new
paths.