Prevent sagan-common from having Spring 3.2.x on its classpath
Spring Social GitHub has a transitive dependency on spring-webmvc
3.2.x. In the absence of any other spring-webmvc dependency, this
causes a mixture of Spring 3.2.x and Spring 4.0.x to be on
sagan-common's classpath.

This commit adds an explicit dependency on spring-webmvc. No version
number is specified to allow the version to be determined by the
Spring Boot plugin.

An alternative approach would be to use a resolution strategy:

configurations.all {
    resolutionStrategy {
        eachDependency {
            if (it.details.requested.group == 'org.springframework') {
                details.useVersion '4.0.5.RELEASE'
            }
        }
    }
}

Unfortunately, this approach requires you to specify the version of
Spring that you want to use, thereby preventing you from automatically
using the version that's provided by the Spring Boot plugin. There's
an issue open (spring-projects/spring-boot#932) to make the versions
available as properties, but, until that functionality is available,
I think the change made in this commit is the best solution.