Log Elastic Search failures
It may be sensible to fail softly when actions against Elastic Search fails for some reason, but not as soft as the current solution. Jest reports errors through `result.isSucceeded()` and `result.getErrorMessage()` (mapping errors and other).

This commit add WARN logging when `result.isSucceeded()` is `false` to make it easier to detect error situations.