Fix json mapping to get github username
When binding the json from github get-single-user call to Map<String, String>, it was
throwing JsonMappingException. This was because the response json contains nested hash
in "plan" field.
In this commit, instaed of binding the entire json to a map, only read the name field from
JsonNode to avoid specific type binding.