Rerender and resummarize all posts
Since the markdown engine changed, its output is different from the
previous one given the same input. In order to visualize all changes in
blog posts, one should be able to trigger a rerender of all posts.

This is a rather CPU/Database intensive operation and should not be
triggered by admin users.