Reduce level of fuzziness in search queries
When using the "AUTO" fuziness level, search fuziness can be quite high
and lead to unwanted results. In fact, the combination of high fuzziness
and reducing words to their root form (for full text search) can lead to
unwanted results.

This commit decreases the level of fuzziness to "1", since 80% of
misspellings have an edit distance of 1.

See
https://www.elastic.co/guide/en/elasticsearch/guide/current/fuzziness.html

Fixes #561