Add project pages in search UI
In a previous commit related to #559, we've added project pages in our
search index.

This commit takes care of the UI side of this feature, with:

* add "Projects/Homepage" as a search facet filter
* make sure it's added as a header facet with "Api" and "Reference"
* add a reference to the current project in the entry mapper to also
add "Projects/ProjectName" as a search facet
* add a "should" search condition to boost project pages entries in
search results

The boost parameter of that "should" condition is now set to "1" but
could be revised in the future for better search results.

Fixes #559