Rewrite search query for better results
Because the most popular search terms did not return good results, the
main search query is rewritten by this commit.

The new search query is a boolean query where results:

* must have most query terms within their title, content or author
fields
* should have a "current" attribute set to true (favors most recent
APIs)
* should be project pages (favors project homepages over other types)
* should math the query as a phrase in its title (favors exact phrases
in titles for common terms such as project names)

In order to improve performance, we can reconsider the use of "should"
sections and use those sub-queries to "rescore" the first results
returned by the main "must" query.