Fix filename snippets CSS in getting started guides
When using an Asciidoctor code block with a title (e.g. a filename), a
CSS negative margin is applied to the code snippet by a custom CSS rule,
making it impossible to read the title of the snippet.

This commit adds some CSS selectors/rules to make the snippet title more
readable.

Fixes #532