Use libs repositories in example pom.xml and build.gradle snippets
Previously, the project pages used http://repo.spring.io/milestone
and http://repo.spring.io/snapshot repository URLs in the sample
pom.xml and build.gradle snippets for milestone and snapshot versions
of a project. This is problematic as it assumes that all of project's
dependencies will be available in the snapshot or milestone
repository. This won't always be the case. For example, current
Spring Boot 1.4 snapshots depend on Spring Data Hopper M1 which isn't
available in the snapshot repository.

This commit updates the project metadata to use
https://repo.spring.io/libs-snapshot and
https://repo.spring.io/libs-milestone. These virtual repositories
make content from the more stable repositories, e.g. libs-snapshot
provides access to the contents of the snapshot, milestone, and
release repositories. This ensures that all of a project's
dependencies are available, without having to configure multiple
repositories.