Switch to a dedicated resource for releases at /{projectId}/releases
If GET /{projectId} returns the whole project it makes for sense that
operations on the releases should be a different resource.

Also adds PUT /{projectId}/releases to enable clients to manage the
whole list in one request (e.g. to add a new one and remove an old
obsolete one).