Add remote renderer service and feature switch to enable it
A blog will be rendered remotely (using a service at
sagan.renderer.uri) if the feature is enabled. The feature
can be enabled by setting a custom header in the request, or
by the authenticated user being in a whitelist
(sagan.feature.users).

If the remote call fails (e.g. there is no remote service) we
fall back to the local renderer. Over time we anticipate that this
code path will be replaced by an error condition (hopefully
preserving the user's input), and Pegdown can be removed from
this project. We would want to test the feature a bit before doing
that.

Also moves renderer service into this repo as a separate module,
individually buildable. You can run the renderer locally on port
9000, or use its stubs with spring cloud CLI:

$ (cd sagan-renderer; gradle install)
$ spring install org.springframework.cloud:spring-cloud-cli:1.4.0.RELEASE
$ spring cloud stubrunner

Then run the main app:

$ (cd sagan-site; SAGAN_FEATURE_ENABLED=true gradle bootRun)