Set up CI/CD with Concourse on ci.spring.io
This commit refactors the experimental concourse pipeline for continuous
deployment on the actual production instance.

The `concourse/pipeline.yml` describes all the jobs required to deploy
Sagan applications. Because we now rely on this pipeline or the `cf` CLI
tool to deploy Sagan apps, the Gradle deployment script is now removed
from the build.