Use Google search API as a replacement for Elasticsearch
Given the low value we get from the search feature, and the fact that
its dependencies are blocking an upgrade of other features, we could
use a different approach.

This is intended to be an interim solution - needs some more tests
and some tidying up. But it works for me.