Add guides rendering infrastructure
This commit adds new endpoints for listing, showing and rendering Spring
Guides. Guides are hosted within a GitHub organization, each as its own
repository.

The list of guides is extracted from the repository list in that GitHub
org, using conventions for repository names (each repository name is
prefixed with the guide type).

The rendering phase of a guide involves downloading the zipball of the
repository and rendering its content to HTML using Asciidoctor.

All web endpoints are exposed as HATEOAS resources.