Change Sagan cache service name
Since https://status.run.pivotal.io/incidents/70b45qggm6x8
there is no way to deploy a new version of the service; Sagan is
deployed with cf concourse resource in a blue/green fashion.

With that ongoing incident, it is not possible to bind a new application
to a redislabs service.

We've created a custom service instance to work around this problem,
with a new name: "sagan-redis". The "sagan-cache" service must be kept
around until this issue is fixed on PWS.