Skip redirected guides in sagan-renderer API
Prior to this commit, fetching a particular guide using the
sagan-renderer API would directly delegate to the GitHub API. When
renaming repositories, GitHub redirects API calls to the newly named
repository. This can cause issues in Sagan where former guides or
tutorials are still displayed.

This commit ensures that the data returned by the GitHub API matches the
expected repository name. If not, we consider that this repository has
been renamed and we should not consider is as an active guide.

See gh-980