Remove Spring Social and Security configuration
Prior to this commit, we would use Spring Social GitHub for the GitHub +
OAuth 2 integration. This is typically used when logging in users in the
admin section of the site.

As this part will be totally rewritten from the now defunct Spring
Social project to using the new Spring Security integration, this commit
deletes completely that part as a first step.