Fix RestDocs tests
This commit fixes the various test issues after the Spring Boot and
Spring RestDocs upgrade.

The first problem is that test classes were improperly using
`@Configuration`-annotated inner-classes, which replaces completely the
configuration detected by Spring Boot. Instead we need to mark them with
`@TestConfiguration` so as to add to the current test context.