Configure OAuth2 security
This commit introduces Spring Security 5 in Sagan site.

Spring Social + a custom auth solution is replaced by
spring-security-oauth2 with its github support for OAuth login.

On top of standard security configuration in`SecurityConfig`, we're
adding a custom `OAuth2UserService` with `GitHubMemberOAuth2UserService`
to check that users logging in the application belong to a specific
GitHub team before they can access the admin section of the website.

`GitHubAuthenticationManager` on the other hand, offers a way to
authenticate with the REST API with a simple OAuth application token.

Co-authored-by: Rob Winch <rwinch@vmware.com>
Co-authored-by: Dave Syer <dsyer@vmware.com>