Rework caching in Sagan site
Prior to this commit, the Sagan site application would define a custom
Redis Template and `CacheManager` implementation to cache various items
in the application: guides, REST response payloads.

This commit now relies on the new Spring Data Redis `CacheManager`
implementation which works with types in a generic way. TTLs
configurations are now moved to configuration properties.

The Google Calendar integration has been refactored to cache the
calendar itself and let the application filter events depending
on the requested period.