Use Liveness probe for platform health checks
As of Spring Boot 2.3, we can use the new Liveness and Readiness
Health probes to let the platform know that the application is up and
ready to receive traffic.

This commit enables those probes and updates the deployment manifest to
tell the platform to use that endpoint instead of the standard health
endpoint, which depends on many external systems and can lead to
cascading failures.