Add Rate-Limit HealthIndicator
The Sagan Renderer application uses the GitHub API to fetch the getting
started guides and render them. While the application should be
configured with an application token, it is still subject to
rate-limiting.

This commit adds a new `HealthIndicator` implementation that checks the
`/"rate_limit"` GitHup API endpoint and returns the current
rate-limit information, exposing it on the `"/actuator/health"`
endpoint.