Fix Release ordering
Prior to this commit, we would order `Release` instances by using the
`Version` comparator (inverse order, putting recent versions first) and
by sorting first `Release` marked as "current".

This commit reverts that last point and does not consider the "current"
flag in the `Release` comparator, as advertizing the current release is
more of a UI concern than a core concern here.