Implement new project support policy
This commit introduces a new `SupportPolicy` concept that must be set on
a per-project basis. Right now this is supporting the UPSTREAM,
SPRING_BOOT and DOWNSTREAM support policies.

Each policy can calculate a support timeline for a ProjectGeneration,
depending on the initial release date of that generation and optionally
the release date of the next generation.

A support timeline is always made of:
* an initial release date for the generation
* a start+end date for the OSS support
* a start+end date for the commercial support