Polish Support policy feature
This commit adds a new link to the project pages, in the support policy
section - pointing to the official OSS support policy page.

This also adds a message in the admin UI, warning project leads if they
ever set a "end of support date" that does not comply with the official
policy.