Use effective OSS support date for calculating commercial support
Prior to this commit, our `SupportPolicy` was always calculating the
commercial support using the official OSS support policy dates.
If the project maintainer provides a custom date for the end of OSS
support (extending the official policy), this date is not currently used
and commercial support does not provide the expected support duration.

This commit refactors the `SupportPolicy` algorithm and gathers all
input into a `SupportTimelineRequest` to better capture all required
information.
This change is also reflected in the "reason" field explaining the
calculated date for each policy.

Fixes gh-1043